+++ src/app/shared-module/tracker/types/tracker-mode.enum.ts
