+++ src/app/shared-module/tracker/types/tracker-settings.ts
